refactor : info 폴더 위치 변경 및 category 타입 수정
